aider: Replaced the `key` prop in the `ProductAccordion` component with a unique value.
# Aider chat conversation:

ASSISTANT: To make the requested change, we need to modify the `src\components\products\productOverviewGallery.tsx` file. Specifically, we need to replace the `key` prop in the `ProductAccordion` component with a unique value.

Here is the *edit block* for the change:

```tsx
src\components\products\productOverviewGallery.tsx
<<<<<<< HEAD
<ProductAccordion data={data} key={''} />
=======
<ProductAccordion data={data} uniqueId={'productOverview'} />
>>>>>>> updated
```

This change will pass a unique id prop to the `ProductAccordion` component, which will be used to create unique ids for the accordion items.